risc-v/esp32c3: Initialize rtc and peripheral parameters by default when chip starts
