+++ arch/risc-v/src/esp32c3/esp32c3_pm.c
@@ -149,38 +149,6 @@ enum esp32c3_sleep_pd_domain_e
   ESP_PD_DOMAIN_MAX              /* Number of domains */
 };
 
-enum esp32c3_periph_module_e
-{
-  PERIPH_LEDC_MODULE = 0,
-  PERIPH_UART0_MODULE,
-  PERIPH_UART1_MODULE,
-  PERIPH_USB_DEVICE_MODULE,
-  PERIPH_I2C0_MODULE,
-  PERIPH_I2S1_MODULE,
-  PERIPH_TIMG0_MODULE,
-  PERIPH_TIMG1_MODULE,
-  PERIPH_UHCI0_MODULE,
-  PERIPH_RMT_MODULE,
-  PERIPH_SPI_MODULE,
-  PERIPH_SPI2_MODULE,
-  PERIPH_TWAI_MODULE,
-  PERIPH_RNG_MODULE,
-  PERIPH_WIFI_MODULE,
-  PERIPH_BT_MODULE,
-  PERIPH_WIFI_BT_COMMON_MODULE,
-  PERIPH_BT_BASEBAND_MODULE,
-  PERIPH_BT_LC_MODULE,
-  PERIPH_RSA_MODULE,
-  PERIPH_AES_MODULE,
-  PERIPH_SHA_MODULE,
-  PERIPH_HMAC_MODULE,
-  PERIPH_DS_MODULE,
-  PERIPH_GDMA_MODULE,
-  PERIPH_SYSTIMER_MODULE,
-  PERIPH_SARADC_MODULE,
-  PERIPH_MODULE_MAX
-};
-
 /* Internal structure which holds all requested deep sleep parameters. */
 
 struct esp32c3_sleep_config_s
@@ -222,16 +190,6 @@ struct esp32c3_rtc_vddsdio_config_s
  * Private Function Prototypes
  ****************************************************************************/
 
-static inline uint32_t IRAM_ATTR esp32c3_periph_ll_get_rst_en_mask
-              (enum esp32c3_periph_module_e periph, bool enable);
-static uint32_t IRAM_ATTR esp32c3_periph_ll_get_rst_en_reg(
-                            enum esp32c3_periph_module_e periph);
-static uint32_t IRAM_ATTR esp32c3_periph_ll_get_clk_en_reg(
-                            enum esp32c3_periph_module_e periph);
-static inline uint32_t IRAM_ATTR esp32c3_periph_ll_get_clk_en_mask(
-                            enum esp32c3_periph_module_e periph);
-static inline bool IRAM_ATTR esp32c3_periph_ll_periph_enabled(
-                                     enum esp32c3_periph_module_e periph);
 static inline void IRAM_ATTR esp32c3_uart_tx_wait_idle(uint8_t uart_no);
 static void IRAM_ATTR esp32c3_flush_uarts(void);
 static void IRAM_ATTR esp32c3_suspend_uarts(void);
@@ -244,9 +202,6 @@ static int IRAM_ATTR esp32c3_get_vddsdio_config(
                       struct esp32c3_rtc_vddsdio_config_s *config);
 static int IRAM_ATTR esp32c3_light_sleep_inner(uint32_t pd_flags,
             uint32_t time_us, struct esp32c3_rtc_vddsdio_config_s config);
-static void esp32c3_periph_module_enable(
-            enum esp32c3_periph_module_e periph);
-static void IRAM_ATTR esp32c3_perip_clk_init(void);
 static int IRAM_ATTR esp32c3_sleep_start(uint32_t pd_flags);
 
 /****************************************************************************
@@ -276,11 +231,6 @@ inform_out_sleep_overhead_cb_t
 
 skip_light_sleep_cb_t g_periph_skip_sleep_cb[PERIPH_SKIP_SLEEP_NO];
 
-static uint8_t ref_counts[PERIPH_MODULE_MAX] =
-{
-  0
-};
-
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
@@ -294,336 +244,6 @@ extern void ets_update_cpu_frequency(uint32_t ticks_per_us);
 /* Pauses execution for us microseconds. */
 
 extern void esp_rom_delay_us(uint32_t us);
-
-/****************************************************************************
- * Name: esp32c3_periph_ll_get_rst_en_mask
- *
- * Description:
- *   Get system reset bit through periph module
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *   enable - Whether hardware acceleration is enabled
- *
- * Returned Value:
- *   System reset bit
- *
- ****************************************************************************/
-
-static inline uint32_t IRAM_ATTR esp32c3_periph_ll_get_rst_en_mask
-              (enum esp32c3_periph_module_e periph, bool enable)
-{
-  switch (periph)
-    {
-      case PERIPH_SARADC_MODULE:
-        return SYSTEM_APB_SARADC_RST;
-
-      case PERIPH_RMT_MODULE:
-        return SYSTEM_RMT_RST;
-
-      case PERIPH_LEDC_MODULE:
-        return SYSTEM_LEDC_RST;
-
-      case PERIPH_UART0_MODULE:
-        return SYSTEM_UART_RST;
-
-      case PERIPH_UART1_MODULE:
-        return SYSTEM_UART1_RST;
-
-      case PERIPH_I2C0_MODULE:
-        return SYSTEM_I2C_EXT0_RST;
-
-      case PERIPH_I2S1_MODULE:
-        return SYSTEM_I2S1_RST;
-
-      case PERIPH_TIMG0_MODULE:
-        return SYSTEM_TIMERGROUP_RST;
-
-      case PERIPH_TIMG1_MODULE:
-        return SYSTEM_TIMERGROUP1_RST;
-
-      case PERIPH_UHCI0_MODULE:
-        return SYSTEM_UHCI0_RST;
-
-      case PERIPH_SYSTIMER_MODULE:
-        return SYSTEM_SYSTIMER_RST;
-
-      case PERIPH_GDMA_MODULE:
-        return SYSTEM_DMA_RST;
-
-      case PERIPH_SPI_MODULE:
-        return SYSTEM_SPI01_RST;
-
-      case PERIPH_SPI2_MODULE:
-        return SYSTEM_SPI2_RST;
-
-      case PERIPH_TWAI_MODULE:
-        return SYSTEM_TWAI_RST;
-
-      case PERIPH_HMAC_MODULE:
-        return SYSTEM_CRYPTO_HMAC_RST;
-
-      case PERIPH_AES_MODULE:
-        if (enable == true)
-          {
-            /* Clear reset on digital signature,
-              * otherwise AES unit is held in reset also.
-              */
-
-            return (SYSTEM_CRYPTO_AES_RST | SYSTEM_CRYPTO_DS_RST);
-          }
-
-        /* Don't return other units to reset,
-         * as this pulls reset on RSA & SHA units, respectively.
-         */
-
-        return SYSTEM_CRYPTO_AES_RST;
-
-      case PERIPH_SHA_MODULE:
-        if (enable == true)
-          {
-            /* Clear reset on digital signature and HMAC,
-             * otherwise SHA is held in reset
-             */
-
-            return (SYSTEM_CRYPTO_SHA_RST |
-                    SYSTEM_CRYPTO_DS_RST | SYSTEM_CRYPTO_HMAC_RST);
-          }
-
-        /* Don't assert reset on secure boot,
-         * otherwise AES is held in reset
-         */
-
-        return SYSTEM_CRYPTO_SHA_RST;
-
-      case PERIPH_RSA_MODULE:
-        if (enable == true)
-          {
-          /* also clear reset on digital signature,
-           * otherwise RSA is held in reset
-           */
-
-            return (SYSTEM_CRYPTO_RSA_RST | SYSTEM_CRYPTO_DS_RST);
-          }
-
-        /* don't reset digital signature unit,
-         * as this resets AES also.
-         */
-
-        return SYSTEM_CRYPTO_RSA_RST;
-
-      case PERIPH_DS_MODULE:
-        return SYSTEM_CRYPTO_DS_RST;
-
-      default:
-        return 0;
-    }
-}
-
-/****************************************************************************
- * Name: esp32c3_periph_ll_get_rst_en_reg
- *
- * Description:
- *   Get system reset register through periph module
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *
- * Returned Value:
- *   System reset register
- *
- ****************************************************************************/
-
-static uint32_t IRAM_ATTR esp32c3_periph_ll_get_rst_en_reg(
-                            enum esp32c3_periph_module_e periph)
-{
-  switch (periph)
-    {
-      case PERIPH_RNG_MODULE:
-      case PERIPH_WIFI_MODULE:
-      case PERIPH_BT_MODULE:
-      case PERIPH_WIFI_BT_COMMON_MODULE:
-      case PERIPH_BT_BASEBAND_MODULE:
-      case PERIPH_BT_LC_MODULE:
-        return SYSTEM_WIFI_RST_EN_REG;
-
-      case PERIPH_HMAC_MODULE:
-      case PERIPH_DS_MODULE:
-      case PERIPH_AES_MODULE:
-      case PERIPH_RSA_MODULE:
-      case PERIPH_SHA_MODULE:
-      case PERIPH_GDMA_MODULE:
-        return SYSTEM_PERIP_RST_EN1_REG;
-
-      default:
-        return SYSTEM_PERIP_RST_EN0_REG;
-    }
-}
-
-/****************************************************************************
- * Name: esp32c3_periph_ll_get_clk_en_reg
- *
- * Description:
- *   Get module clock register through periph module
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *
- * Returned Value:
- *   Module clock register
- *
- ****************************************************************************/
-
-static uint32_t IRAM_ATTR esp32c3_periph_ll_get_clk_en_reg(
-                             enum esp32c3_periph_module_e periph)
-{
-  switch (periph)
-    {
-      case PERIPH_RNG_MODULE:
-      case PERIPH_WIFI_MODULE:
-      case PERIPH_BT_MODULE:
-      case PERIPH_WIFI_BT_COMMON_MODULE:
-      case PERIPH_BT_BASEBAND_MODULE:
-      case PERIPH_BT_LC_MODULE:
-        return SYSTEM_WIFI_CLK_EN_REG;
-
-      case PERIPH_HMAC_MODULE:
-      case PERIPH_DS_MODULE:
-      case PERIPH_AES_MODULE:
-      case PERIPH_RSA_MODULE:
-      case PERIPH_SHA_MODULE:
-      case PERIPH_GDMA_MODULE:
-        return SYSTEM_PERIP_CLK_EN1_REG;
-
-      default:
-        return SYSTEM_PERIP_CLK_EN0_REG;
-    }
-}
-
-/****************************************************************************
- * Name: esp32c3_periph_ll_get_clk_en_mask
- *
- * Description:
- *   Get module clock bit through periph module
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *
- * Returned Value:
- *   Module clock bit
- *
- ****************************************************************************/
-
-static inline uint32_t IRAM_ATTR esp32c3_periph_ll_get_clk_en_mask(
-                               enum esp32c3_periph_module_e periph)
-{
-  switch (periph)
-    {
-      case PERIPH_SARADC_MODULE:
-        return SYSTEM_APB_SARADC_CLK_EN;
-
-      case PERIPH_RMT_MODULE:
-        return SYSTEM_RMT_CLK_EN;
-
-      case PERIPH_LEDC_MODULE:
-        return SYSTEM_LEDC_CLK_EN;
-
-      case PERIPH_UART0_MODULE:
-        return SYSTEM_UART_CLK_EN;
-
-      case PERIPH_UART1_MODULE:
-        return SYSTEM_UART1_CLK_EN;
-
-      case PERIPH_I2C0_MODULE:
-        return SYSTEM_I2C_EXT0_CLK_EN;
-
-      case PERIPH_I2S1_MODULE:
-        return SYSTEM_I2S1_CLK_EN;
-
-      case PERIPH_TIMG0_MODULE:
-        return SYSTEM_TIMERGROUP_CLK_EN;
-
-      case PERIPH_TIMG1_MODULE:
-        return SYSTEM_TIMERGROUP1_CLK_EN;
-
-      case PERIPH_UHCI0_MODULE:
-        return SYSTEM_UHCI0_CLK_EN;
-
-      case PERIPH_SYSTIMER_MODULE:
-        return SYSTEM_SYSTIMER_CLK_EN;
-
-      case PERIPH_SPI_MODULE:
-        return SYSTEM_SPI01_CLK_EN;
-
-      case PERIPH_SPI2_MODULE:
-        return SYSTEM_SPI2_CLK_EN;
-
-      case PERIPH_TWAI_MODULE:
-        return SYSTEM_TWAI_CLK_EN;
-
-      case PERIPH_GDMA_MODULE:
-        return SYSTEM_DMA_CLK_EN;
-
-      case PERIPH_AES_MODULE:
-        return SYSTEM_CRYPTO_AES_CLK_EN;
-
-      case PERIPH_SHA_MODULE:
-        return SYSTEM_CRYPTO_SHA_CLK_EN;
-
-      case PERIPH_RSA_MODULE:
-        return SYSTEM_CRYPTO_RSA_CLK_EN;
-
-      case PERIPH_HMAC_MODULE:
-        return SYSTEM_CRYPTO_HMAC_CLK_EN;
-
-      case PERIPH_DS_MODULE:
-        return SYSTEM_CRYPTO_DS_CLK_EN;
-
-      case PERIPH_RNG_MODULE:
-        return SYSTEM_WIFI_CLK_RNG_EN;
-
-      case PERIPH_WIFI_MODULE:
-        return SYSTEM_WIFI_CLK_WIFI_EN_M;
-
-      case PERIPH_BT_MODULE:
-        return SYSTEM_WIFI_CLK_BT_EN_M;
-
-      case PERIPH_WIFI_BT_COMMON_MODULE:
-        return SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M;
-
-      case PERIPH_BT_BASEBAND_MODULE:
-        return SYSTEM_BT_BASEBAND_EN;
-
-      case PERIPH_BT_LC_MODULE:
-        return SYSTEM_BT_LC_EN;
-
-      default:
-        return 0;
-    }
-}
-
-/****************************************************************************
- * Name: esp32c3_periph_ll_periph_enabled
- *
- * Description:
- *   Whether the current Periph module is enabled
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *
- * Returned Value:
- *   Periph module is enabled or not
- *
- ****************************************************************************/
-
-static inline bool IRAM_ATTR esp32c3_periph_ll_periph_enabled(
-                                     enum esp32c3_periph_module_e periph)
-{
-  return ((getreg32(esp32c3_periph_ll_get_rst_en_reg(periph)) &
-           esp32c3_periph_ll_get_rst_en_mask(periph, false)) == 0) &&
-         ((getreg32(esp32c3_periph_ll_get_clk_en_reg(periph)) &
-           esp32c3_periph_ll_get_clk_en_mask(periph)) != 0);
-}
 
 /****************************************************************************
  * Name: esp32c3_uart_tx_wait_idle
@@ -1053,116 +673,6 @@ static int IRAM_ATTR esp32c3_light_sleep_inner(uint32_t pd_flags,
 }
 
 /****************************************************************************
- * Name: esp32c3_periph_module_enable
- *
- * Description:
- *   Enable peripheral module
- *
- * Input Parameters:
- *   periph - Periph module (one of enum esp32c3_periph_module_e values)
- *
- * Returned Value:
- *   None
- *
- ****************************************************************************/
-
-static void esp32c3_periph_module_enable(enum esp32c3_periph_module_e periph)
-{
-  irqstate_t flags = enter_critical_section();
-
-  assert(periph < PERIPH_MODULE_MAX);
-  if (ref_counts[periph] == 0)
-    {
-      modifyreg32(esp32c3_periph_ll_get_clk_en_reg(periph), 0,
-                  esp32c3_periph_ll_get_clk_en_mask(periph));
-      modifyreg32(esp32c3_periph_ll_get_rst_en_reg(periph),
-                  esp32c3_periph_ll_get_rst_en_mask(periph, true), 0);
-    }
-
-  ref_counts[periph]++;
-
-  leave_critical_section(flags);
-}
-
-/****************************************************************************
- * Name:  esp32c3_perip_clk_init
- *
- * Description:
- *   This function disables clock of useless peripherals when cpu starts.
- *
- * Input Parameters:
- *   None
- *
- * Returned Value:
- *   None
- *
- ****************************************************************************/
-
-static void IRAM_ATTR esp32c3_perip_clk_init(void)
-{
-  uint32_t common_perip_clk1 = 0;
-
-  /* Reset the communication peripherals like I2C, SPI,
-   * UART, I2S and bring them to known state.
-   */
-
-  uint32_t common_perip_clk = SYSTEM_WDG_CLK_EN |
-                           SYSTEM_I2S0_CLK_EN |
-                           SYSTEM_UART1_CLK_EN |
-                           SYSTEM_UART2_CLK_EN |
-                           SYSTEM_SPI2_CLK_EN |
-                           SYSTEM_I2C_EXT0_CLK_EN |
-                           SYSTEM_UHCI0_CLK_EN |
-                           SYSTEM_RMT_CLK_EN |
-                           SYSTEM_LEDC_CLK_EN |
-                           SYSTEM_UHCI1_CLK_EN |
-                           SYSTEM_TIMERGROUP1_CLK_EN |
-                           SYSTEM_SPI3_CLK_EN |
-                           SYSTEM_SPI4_CLK_EN |
-                           SYSTEM_I2C_EXT1_CLK_EN |
-                           SYSTEM_TWAI_CLK_EN |
-                           SYSTEM_I2S1_CLK_EN |
-                           SYSTEM_SPI2_DMA_CLK_EN |
-                           SYSTEM_SPI3_DMA_CLK_EN;
-
-  uint32_t hwcrypto_perip_clk = SYSTEM_CRYPTO_AES_CLK_EN |
-                          SYSTEM_CRYPTO_SHA_CLK_EN |
-                          SYSTEM_CRYPTO_RSA_CLK_EN;
-  uint32_t wifi_bt_sdio_clk = SYSTEM_WIFI_CLK_WIFI_EN |
-                        SYSTEM_WIFI_CLK_BT_EN_M |
-                        SYSTEM_WIFI_CLK_UNUSED_BIT5 |
-                        SYSTEM_WIFI_CLK_UNUSED_BIT12;
-
-  /* Disable some peripheral clocks. */
-
-  modifyreg32(SYSTEM_PERIP_CLK_EN0_REG, common_perip_clk, 0);
-  modifyreg32(SYSTEM_PERIP_RST_EN0_REG, 0, common_perip_clk);
-
-  modifyreg32(SYSTEM_PERIP_CLK_EN1_REG, common_perip_clk1, 0);
-  modifyreg32(SYSTEM_PERIP_RST_EN1_REG, 0, common_perip_clk1);
-
-  /* Disable hardware crypto clocks. */
-
-  modifyreg32(SYSTEM_PERIP_CLK_EN1_REG, hwcrypto_perip_clk, 0);
-  modifyreg32(SYSTEM_PERIP_RST_EN1_REG, 0, hwcrypto_perip_clk);
-
-  /* Disable Wi-Fi/BT/SDIO clocks. */
-
-  modifyreg32(SYSTEM_WIFI_CLK_EN_REG, wifi_bt_sdio_clk, 0);
-  modifyreg32(SYSTEM_WIFI_CLK_EN_REG, 0, SYSTEM_WIFI_CLK_EN);
-
-  /* Set Wi-Fi light sleep clock source to RTC slow clock */
-
-  REG_SET_FIELD(SYSTEM_BT_LPCK_DIV_INT_REG, SYSTEM_BT_LPCK_DIV_NUM, 0);
-  modifyreg32(SYSTEM_BT_LPCK_DIV_FRAC_REG, SYSTEM_LPCLK_SEL_8M,
-              SYSTEM_LPCLK_SEL_RTC_SLOW);
-
-  /* Enable RNG clock. */
-
-  esp32c3_periph_module_enable(PERIPH_RNG_MODULE);
-}
-
-/****************************************************************************
  * Name:  esp32c3_periph_should_skip_sleep
  *
  * Description:
@@ -1529,29 +1039,6 @@ int IRAM_ATTR esp32c3_light_sleep_start(uint64_t *sleep_time)
   leave_critical_section(flags);
 
   return ret;
-}
-
-/****************************************************************************
- * Name: esp32c3_pminit
- *
- * Description:
- *   Initialize force sleep parameters.
- *
- * Input Parameters:
- *   None
- *
- * Returned Value:
- *   None
- *
- ****************************************************************************/
-
-void esp32c3_pminit(void)
-{
-  /* Initialize RTC parameters */
-
-  esp32c3_rtc_init();
-  esp32c3_rtc_clk_set();
-  esp32c3_perip_clk_init();
 }
 
 /****************************************************************************
