+++ arch/risc-v/src/esp32c3/esp32c3_start.c
@@ -35,6 +35,7 @@
 #include "esp32c3_lowputc.h"
 #include "esp32c3_start.h"
 #include "esp32c3_wdt.h"
+#include "esp32c3_rtc.h"
 #include "hardware/esp32c3_cache_memory.h"
 #include "hardware/extmem_reg.h"
 
@@ -240,11 +241,21 @@ void __esp32c3_start(void)
 
 #endif
 
+  /* Initialize RTC parameters */
+
+  esp32c3_rtc_init();
+  esp32c3_rtc_clk_set();
+
   /* Set CPU frequency */
 
   esp32c3_clockconfig();
 
+  /* Initialize peripherals parameters */
+
+  esp32c3_perip_clk_init();
+
 #ifdef CONFIG_ESP32C3_BROWNOUT_DET
+
   /* Initialize hardware brownout check and reset */
 
   esp32c3_brownout_init();
